Add dynamic header color
